[feature]1.split dataset by every user when prepare;2. save ranking result when evaluate

+++ recsys/pop/pop_predict.h
@@ -48,7 +48,7 @@ private:
     const Pop::PredictOption *mPredictOption = nullptr;
     std::vector<float> mUserAverages;
     std::vector<float> mItemAverages;
-    std::vector<float> mItemPopularities;
+    std::vector<int> mItemPopularities;
 };
 
 } //~ namespace longan
